+++ addresses/index.js
@@ -1,0 +1,4 @@
+const decodeSocket = require('./decode_socket');
+const encodeSocket = require('./encode_socket');
+
+module.exports = {decodeSocket, encodeSocket};