+++ index.js
@@ -1,7 +1,9 @@
 const {chanFormat} = require('./ids');
 const {chanNumber} = require('./ids');
 const {decodeChanId} = require('./ids');
+const {decodeSocket} = require('./addresses');
 const {encodeChanId} = require('./ids');
+const {encodeSocket} = require('./addresses');
 const {hopsFromChannels} = require('./routing');
 const {rawChanId} = require('./ids');
 const {routeFromChannels} = require('./routing');
@@ -11,7 +13,9 @@ module.exports = {
   chanFormat,
   chanNumber,
   decodeChanId,
+  decodeSocket,
   encodeChanId,
+  encodeSocket,
   hopsFromChannels,
   rawChanId,
   routeFromChannels,